requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-JWT-Extended
Flask-CORS
Flask-MongoEngine # Or pymongo
psycopg2-binary   # PostgreSQL driver
Werkzeug          # For password hashing (often installed with Flask)
python-dotenv     # For loading .env file locally